requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn==21.2.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0
python-dateutil>=2.8.2
pytz>=2023.3
anyio>=3.6.2
click>=8.1.3
h11>=0.14.0
idna>=3.4
sniffio>=1.3.0
typing-extensions>=4.5.0